Dark mode implemented
User now can go to the Profile Screen and then click on User "Edit Profile" and can select the dark mode. It has been applied to all the views except the main one, it saves the user preferences, so if that user logs back in it would be saved on the mode selected.
FIX: Now if the user selects a new, will be able to see the new on the mode that prefers instead of the automatically turning black