News History Implemented
Now, the user can go to the Profile Screen and check the news that have read so far. Also, in there, the user can select any new from his history in order to read it again, check the gradings and the comments. There's also an option that allows the user to delete the history in case that wants to get rid off it.
FIX: There was a small bug showing the Category id instead of the Category name when looking at the history, this has been resolved. Now the user see the name of the category