Bookmark Implemented
Now the user can save the news that likes the most. Just goes to the ProfileScreen and will see a box that shows up how many Bookmarks the user have.
UPDATE: On the tab section of the ProfileScreen there was a duplicate of the Bookmarks, so it has been eliminated. At first the bookmarks were visible only in there, but due user navigation, we rather keep it as the box.
FIX: Whenever the user wanted to access to the Bookmarks the app stops, it has been resolved.